feat: refactor the model and move the creation and connection of the db to a utility
